requirements.txt
paho_mqtt==1.6.1
pymongo==4.6.1
